Update README.md
The scope of this app is not yet defined, it's intented to reach a lot of different fields. But since I'm a solo developer so far, I'll focus the functionality of this app to many business. Something like a Business Finder, planning to be a "Finder" for way more areas.